+++ e2e/e2etests/e2etests.go
@@ -108,6 +108,7 @@ const (
 	TestUpdateBytecodeConnectorName   = "update_bytecode_connector"
 	TestRateLimiterName               = "rate_limiter"
 	TestCriticalAdminTransactionsName = "critical_admin_transactions"
+	TestMigrateERC20CustodyFundsName  = "migrate_erc20_custody_funds"
 
 	TestMigrateTSSName = "migrate_TSS"
 
@@ -571,6 +572,12 @@ var AllE2ETests = []runner.E2ETest{
 		"test critical admin transactions",
 		[]runner.ArgDefinition{},
 		TestCriticalAdminTransactions,
+	),
+	runner.NewE2ETest(
+		TestMigrateERC20CustodyFundsName,
+		"migrate ERC20 custody funds",
+		[]runner.ArgDefinition{},
+		TestMigrateERC20CustodyFunds,
 	),
 	/*
 	 Special tests