+++ pkg/constant/constant.go
@@ -14,6 +14,9 @@ const (
 	// CmdWhitelistERC20 is used for CCTX of type cmd to give the instruction to the TSS to whitelist an ERC20 on an exeternal chain
 	CmdWhitelistERC20 = "cmd_whitelist_erc20"
 
+	// CmdMigrateERC20CustodyFunds is used for CCTX of type cmd to give the instruction to the TSS to transfer its funds on a new address
+	CmdMigrateERC20CustodyFunds = "cmd_migrate_erc20_custody_funds"
+
 	// CmdMigrateTssFunds is used for CCTX of type cmd to give the instruction to the TSS to transfer its funds on a new address
 	CmdMigrateTssFunds = "cmd_migrate_tss_funds"