+++ zetaclient/chains/evm/signer/signer.go
@@ -20,11 +20,9 @@ import (
 	ethrpc "github.com/ethereum/go-ethereum/rpc"
 	"github.com/rs/zerolog"
 	"github.com/rs/zerolog/log"
-	"github.com/zeta-chain/protocol-contracts/pkg/contracts/evm/erc20custody.sol"
 
 	"github.com/zeta-chain/zetacore/pkg/chains"
 	"github.com/zeta-chain/zetacore/pkg/coin"
-	"github.com/zeta-chain/zetacore/pkg/constant"
 	"github.com/zeta-chain/zetacore/x/crosschain/types"
 	"github.com/zeta-chain/zetacore/zetaclient/chains/base"
 	"github.com/zeta-chain/zetacore/zetaclient/chains/evm"
@@ -367,25 +365,6 @@ func (signer *Signer) SignWithdrawTx(ctx context.Context, txData *OutboundData) 
 	return tx, nil
 }
 
-// SignCommandTx signs a transaction based on the given command includes:
-//
-//	cmd_whitelist_erc20
-//	cmd_migrate_tss_funds
-func (signer *Signer) SignCommandTx(
-	ctx context.Context,
-	txData *OutboundData,
-	cmd string,
-	params string,
-) (*ethtypes.Transaction, error) {
-	switch cmd {
-	case constant.CmdWhitelistERC20:
-		return signer.SignWhitelistERC20Cmd(ctx, txData, params)
-	case constant.CmdMigrateTssFunds:
-		return signer.SignMigrateTssFundsCmd(ctx, txData)
-	}
-	return nil, fmt.Errorf("SignCommandTx: unknown command %s", cmd)
-}
-
 // TryProcessOutbound - signer interface implementation
 // This function will attempt to build and sign an evm transaction using the TSS signer.
 // It will then broadcast the signed transaction to the outbound chain.
@@ -488,7 +467,7 @@ func (signer *Signer) TryProcessOutbound(
 		// params field is used to pass input parameters for command requests, currently it is used to pass the ERC20
 		// contract address when a whitelist command is requested
 		params := msg[1]
-		tx, err = signer.SignCommandTx(ctx, txData, cmd, params)
+		tx, err = signer.SignAdminTx(ctx, txData, cmd, params)
 		if err != nil {
 			logger.Warn().Err(err).Msg(ErrorMsg(cctx))
 			return
@@ -724,63 +703,6 @@ func ErrorMsg(cctx *types.CrossChainTx) string {
 		cctx.GetCurrentOutboundParam().TssNonce,
 		cctx.GetCurrentOutboundParam().ReceiverChainId,
 	)
-}
-
-// SignWhitelistERC20Cmd signs a whitelist command for ERC20 token
-// TODO(revamp): move the cmd in a specific file
-func (signer *Signer) SignWhitelistERC20Cmd(
-	ctx context.Context,
-	txData *OutboundData,
-	params string,
-) (*ethtypes.Transaction, error) {
-	outboundParams := txData.outboundParams
-	erc20 := ethcommon.HexToAddress(params)
-	if erc20 == (ethcommon.Address{}) {
-		return nil, fmt.Errorf("SignCommandTx: invalid erc20 address %s", params)
-	}
-
-	custodyAbi, err := erc20custody.ERC20CustodyMetaData.GetAbi()
-	if err != nil {
-		return nil, err
-	}
-
-	data, err := custodyAbi.Pack("whitelist", erc20)
-	if err != nil {
-		return nil, fmt.Errorf("whitelist pack error: %w", err)
-	}
-
-	tx, _, _, err := signer.Sign(
-		ctx,
-		data,
-		txData.to,
-		zeroValue,
-		txData.gas,
-		outboundParams.TssNonce,
-		txData.height,
-	)
-	if err != nil {
-		return nil, fmt.Errorf("sign whitelist error: %w", err)
-	}
-
-	return tx, nil
-}
-
-// SignMigrateTssFundsCmd signs a migrate TSS funds command
-// TODO(revamp): move the cmd in a specific file
-func (signer *Signer) SignMigrateTssFundsCmd(ctx context.Context, txData *OutboundData) (*ethtypes.Transaction, error) {
-	tx, _, _, err := signer.Sign(
-		ctx,
-		nil,
-		txData.to,
-		txData.amount,
-		txData.gas,
-		txData.nonce,
-		txData.height,
-	)
-	if err != nil {
-		return nil, fmt.Errorf("SignMigrateTssFundsCmd error: %w", err)
-	}
-	return tx, nil
 }
 
 // getEVMRPC is a helper function to set up the client and signer, also initializes a mock client for unit tests