+++ changelog.md
@@ -4,6 +4,8 @@
 
 ### Features
 
+* [2578](https://github.com/zeta-chain/node/pull/2578) - add Gateway address in protocol contract list
+* [2630](https://github.com/zeta-chain/node/pull/2630) - implement `MsgMigrateERC20CustodyFunds` to migrate the funds from the ERC20Custody to a new contracts (to be used for the new ERC20Custody contract for smart contract V2)
 * [2578](https://github.com/zeta-chain/node/pull/2578) - Add Gateway address in protocol contract list
 * [2634](https://github.com/zeta-chain/node/pull/2634) - add support for EIP-1559 gas fees
 * [2597](https://github.com/zeta-chain/node/pull/2597) - Add generic rpc metrics to zetaclient